fix precision

+++ EigenvectorCentrality/src/main/scala/org/ymt/spark/graphx/eigenvector/EigenvectorCentrality.scala
@@ -58,7 +58,7 @@ object EigenvectorCentrality extends Serializable{
     }
 
     val count = graph.numVertices
-    var result = graph.mapVertices((vid, attr) => (1 / count).toDouble)
+    var result = graph.mapVertices((vid, attr) => 1.0 / count)
     var initialGraph: Graph[Double, Double] = result
     var condition = Double.MaxValue
     @transient lazy val log = LogManager.getRootLogger
